refactor: remove legacy orchestration
