--- .github/workflows/apply-queue.yml
@@ -1,26 +1,0 @@
-name: Apply Block-Change Queue
-
-on:
-  schedule:
-    # 03:15 UTC every day
-    - cron: '15 3 * * *'
-  workflow_dispatch: {}
-
-jobs:
-  apply-block-queue:
-    runs-on: ubuntu-latest
-    env:
-      # make sure these are set in repo → Settings → Secrets
-      DATABASE_URL: ${{ secrets.DATABASE_URL }}
-      PYTHONPATH: "api/src"
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.10'
-      - name: Install deps (no openai-agents)
-        run: |
-          pip install -r api/requirements.txt || true
-          pip install asyncpg
-      - name: Run worker
-        run: python api/src/app/orchestration/apply_queue_worker.py
