+++ api/src/app/agent_tasks/orch/__init__.py
@@ -1,2 +1,13 @@
 """Orchestration-level (orch) agent tasks package."""
-__all__ = []
+
+from .orch_block_diff_agent import run as diff_blocks
+from .orch_basket_parser_agent import run as parse_basket
+from .apply_diff_blocks import apply_diffs
+from .orchestration_runner import run_all
+
+__all__ = [
+    "diff_blocks",
+    "parse_basket",
+    "apply_diffs",
+    "run_all",
+]
