+++ api/src/app/agent_tasks/orch/orchestration_runner.py
@@ -1,4 +1,4 @@
-# api/src/app/agent_tasks/orchestration/orchestration_runner.py
+# api/src/app/agent_tasks/orch/orchestration_runner.py
 # (When you call the nightly cron script, it will now hang “forever” listening for events; that is fine on Render or a worker dyno. 
 # For GitHub Actions you might instead want a 5-minute timeout.)
 
@@ -6,8 +6,6 @@ import asyncio
 from app.agent_tasks.layer1_infra.agents.infra_analyzer_agent import run as run_analyzer
 from app.agent_tasks.layer1_infra.agents.infra_observer_agent import run as run_observer
 from app.agent_tasks.orchestration.orch_block_manager_agent import run as run_block_manager
-from app.agent_tasks.orchestration.config_listener import run as run_config_listener
-from app.agent_tasks.orchestration.thread_parser_listener import run as run_thread_parser
 from app.agent_tasks.orchestration.orch_basket_composer_agent import run as run_basket_composer
 
 async def run_all():
@@ -17,8 +15,6 @@ async def run_all():
     # start the manager listener and config listener concurrently
     await asyncio.gather(
         run_block_manager(),
-        run_config_listener(),
-        run_thread_parser(),
         run_basket_composer(),
     )
