+++ scripts/run_orchestration.py
@@ -1,5 +1,5 @@
 import asyncio
-from agent_tasks.orchestration.orchestration_runner import run_all
+from agent_tasks.orch.orchestration_runner import run_all
 
 if __name__ == "__main__":
     asyncio.run(run_all())   # later you’ll fill this in
